CloudFunctions/requirements.txt
google-cloud-storage
gdown